requirements.txt
requests>=2.31.0
flask>=2.3.3
plotly>=5.17.0
python-dotenv>=1.0.0
pandas>=2.1.3
matplotlib>=3.7.0
numpy>=1.24.0
APScheduler
finnhub-python 
alpha-vantage 